subindo dados

+++ backend/apiCOMPRAS/src/main/java/com/bahiavisual/apiCOMPRAS/service/GrupoService.java
@@ -1,7 +1,9 @@
 package com.bahiavisual.apiCOMPRAS.service;
 
 import com.bahiavisual.apiCOMPRAS.entity.GrupoProduto;
+import com.bahiavisual.apiCOMPRAS.entity.Produtos;
 import com.bahiavisual.apiCOMPRAS.repository.GrupoRepository;
+import com.bahiavisual.apiCOMPRAS.repository.ProdutosRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +16,9 @@ import java.util.Optional;
 public class GrupoService {
     @Autowired
     GrupoRepository grupoRepository;
+
+    @Autowired
+    ProdutosRepository produtosRepository;
 
     public List<GrupoProduto> buscarTodosItens() {
         List<GrupoProduto> grupoProdutoList = grupoRepository.findAll();
@@ -37,6 +42,10 @@ public class GrupoService {
         Optional<GrupoProduto> dbNome = grupoRepository.findByNome(grupoProduto.getNome());
         if (dbNome.isEmpty()){
             return new ResponseEntity("GRUPO NÃO CADASTRADO NO BANCO DE DADOS", HttpStatus.NOT_FOUND);
+        }
+        Optional<Produtos> dbProdutoGrupo = produtosRepository.findByGrupoProduto(dbNome.get());
+        if (dbProdutoGrupo.isPresent()){
+            return new ResponseEntity("O GRUPO NÃO PODE SER EXCLUIDO POIS JÁ ESTÁ VINCULADO A UM OU MAIS PRODUTOS", HttpStatus.BAD_REQUEST);
         }
         grupoRepository.deleteById(dbNome.get().getId());
         return new ResponseEntity("GRUPO " + dbNome.get().getNome() +  " DELETADO COM SUCESSO", HttpStatus.OK);
